pyproject.toml
[project]
name = "ncr-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faicons==0.2.2",
    "folium==0.16.0",
    "leafmap==0.31.9",
    "pandas==2.2.2",
    "plotly==5.22.0",
    "pytest==8.2.2",
    "python-dotenv==1.2.1",
    "requests==2.31.0",
    "rsconnect-python==1.24.0",
    "setuptools>=80.9.0",
    "shiny==0.9.0",
    "shinywidgets==0.3.2",
]
